refactor: update to use typscript project references
Updating the packages to use TypeScript project references support
allows for tooling such as go to definition & refactoring. Also fixed
some build warnings, etc.
