+++ .gitignore
@@ -5,6 +5,7 @@ dist
 build
 tmp
 out-tsc
+*.tsbuildinfo
 
 # dependencies
 node_modules
