+++ package.json
@@ -12,13 +12,13 @@
   "scripts": {
     "build": "lerna run build",
     "build:prod": "lerna run build:prod",
-    "build:watch": "lerna run build:watch",
+    "build:watch": "lerna run --parallel --no-bail build:watch",
     "clean": "lerna run clean",
     "commit": "commit",
     "release": "npm run build:prod && lerna publish",
     "changelog": "conventional-changelog -i CHANGELOG.md -s -r 0 && prettier --write CHANGELOG.md",
     "test": "lerna run test",
-    "test:watch": "lerna run test:watch",
+    "test:watch": "lerna run --parallel --no-bail test:watch",
     "coverage": "lerna run coverage",
     "lint": "lerna run lint",
     "lint:fix": "lerna run lint:fix",
@@ -28,13 +28,10 @@
   },
   "dependencies": {},
   "devDependencies": {
-    "@angular-devkit/core": "~7.3.7",
-    "@angular-devkit/schematics": "~7.3.7",
     "@commitlint/cli": "^7.5.2",
     "@commitlint/config-conventional": "^7.5.0",
     "@commitlint/config-lerna-scopes": "^7.5.1",
     "@commitlint/prompt-cli": "^7.5.0",
-    "@schematics/angular": "^7.3.7",
     "@types/jest": "^24.0.11",
     "clean-webpack-plugin": "^1.0.0",
     "conventional-changelog": "^3.1.3",
@@ -42,6 +39,7 @@
     "copy-webpack-plugin": "^4.6.0",
     "cross-env": "^5.2.0",
     "css-loader": "^2.1.1",
+    "html-webpack-plugin": "^3.2.0",
     "husky": "^1.3.1",
     "jest": "^24.7.1",
     "lerna": "^3.13.2",
