+++ packages/core/webpack.config.js
@@ -1,5 +1,5 @@
 const merge = require('webpack-merge');
 
-const baseWebpackConfig = require('../../webpack.config.base');
+const baseWebpackConfig = require('../webpack.config.base');
 
 module.exports = merge(baseWebpackConfig);
