+++ packages/frontend/src/index.html
@@ -5,12 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Augury - Performance Profiler</title>
     <base href="/">
-
     <link rel="icon" type="image/x-icon" href="favicon.ico">
   </head>
   <body>
     <app-root></app-root>
-    <script src='polyfills.js'></script>
-    <script src='main.js'></script>
   </body>
 </html>
