+++ packages/frontend/tsconfig.json
@@ -11,5 +11,6 @@
     "experimentalDecorators": true,
     "sourceMap": true
   },
-  "exclude": ["node_modules"]
+  "exclude": ["node_modules"],
+  "references": [{ "path": "../core" }]
 }
