+++ packages/performance-profiler-plugin/tsconfig.json
@@ -1,8 +1,11 @@
 {
-  "extends": "../../tsconfig.base.json",
+  "extends": "../tsconfig.base.json",
   "compilerOptions": {
+    "rootDir": "./src",
     "outDir": "./dist",
     "declarationDir": "./dist/typings"
   },
-  "files": ["./src/index.ts"]
+  "include": ["src"],
+  "exclude": ["**/*.spec.ts"],
+  "references": [{ "path": "../core" }]
 }
