+++ packages/tsconfig.base.json
@@ -1,13 +1,19 @@
 {
   "compilerOptions": {
     "baseUrl": ".",
+    "rootDir": ".",
     "target": "es2015",
     "lib": ["es2018", "dom"],
     "module": "es2015",
     "moduleResolution": "node",
+    "composite": true,
     "declaration": true,
-    "noImplicitAny": false,
+    "declarationMap": true,
+    "sourceMap": true,
     "strict": true,
-    "sourceMap": true
+    "noImplicitAny": false,
+    "paths": {
+      "@augury/*": ["./*/src"]
+    }
   }
 }
