+++ packages/unit-tester-plugin/webpack.config.js
@@ -1,5 +1,5 @@
 const merge = require('webpack-merge');
 
-const webpackConfigBase = require('../../webpack.config.base');
+const webpackConfigBase = require('../webpack.config.base');
 
 module.exports = merge(webpackConfigBase);
