+++ packages/webpack.config.base.js
